# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Evaluation Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Project {
  id: number;
  title: string;
  description: string;
  category: string;
  status: ProjectStatus;
  isFeatured: boolean;
  /** @nullable */
  imageUrl: string | null;
  voteCount: number;
  userId: number;
  ownerName: string;
  createdAt: string;
}

export type ProjectDetailStatus =
  (typeof ProjectDetailStatus)[keyof typeof ProjectDetailStatus];

export const ProjectDetailStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface ProjectDetail {
  id: number;
  title: string;
  description: string;
  category: string;
  status: ProjectDetailStatus;
  isFeatured: boolean;
  /** @nullable */
  imageUrl: string | null;
  voteCount: number;
  userId: number;
  ownerName: string;
  createdAt: string;
}

export interface CreateProjectBody {
  title: string;
  description: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface PaginatedProjects {
  projects: Project[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface AdminPaginatedProjects {
  projects: Project[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface PaginatedUsers {
  users: User[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface VoteResponse {
  voted: boolean;
  voteCount: number;
}

export interface MyVoteResponse {
  voted: boolean;
}

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface Payment {
  id: number;
  userId: number;
  projectId: number;
  amount: number;
  status: PaymentStatus;
  /** @nullable */
  stripeSessionId: string | null;
  /** @nullable */
  projectTitle: string | null;
  /** @nullable */
  userEmail: string | null;
  createdAt: string;
}

export interface CreateCheckoutBody {
  projectId: number;
}

export interface CheckoutResponse {
  url: string;
}

export interface Analytics {
  totalUsers: number;
  totalProjects: number;
  totalVotes: number;
  totalRevenue: number;
  pendingProjects: number;
  approvedProjects: number;
  rejectedProjects: number;
  featuredProjects: number;
}

export interface CategoryStat {
  category: string;
  count: number;
}

export type ListProjectsParams = {
  page?: number;
  limit?: number;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  category?: string | null;
  sort?: ListProjectsSort;
};

export type ListProjectsSort =
  (typeof ListProjectsSort)[keyof typeof ListProjectsSort];

export const ListProjectsSort = {
  newest: "newest",
  votes: "votes",
  featured: "featured",
} as const;

export type AdminListUsersParams = {
  page?: number;
  limit?: number;
};

export type AdminListProjectsParams = {
  page?: number;
  limit?: number;
  /**
   * @nullable
   */
  status?: AdminListProjectsStatus;
};

export type AdminListProjectsStatus =
  | (typeof AdminListProjectsStatus)[keyof typeof AdminListProjectsStatus]
  | null;

export const AdminListProjectsStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type GetTopVotedProjectsParams = {
  limit?: number;
};